feat: rename all routes
