+++ frontend/src/App.tsx
@@ -1,11 +1,11 @@
 import { Outlet, Route, Routes, useNavigate } from "react-router-dom";
 import MainLayout from "@/layout/MainLayout";
 import ReportPage from "@/pages/app/report";
-import { LoginPage } from "@/pages/login";
-import { RegisterPage } from "./pages/register";
+import { LoginPage } from "@/pages/LoginPage.tsx";
+import { RegisterPage } from "./pages/RegisterPage.tsx";
 import { ErrorScreen } from "./components/ui/ErrorScreen";
 import { Box, Button } from "@mantine/core";
-import UsersPage from "./pages/app/users";
+import UsersPage from "./pages/app/UsersPage.tsx";
 import LiveCamPage from "./pages/app/LiveCamPage";
 
 function App() {
@@ -18,7 +18,7 @@ function App() {
         <Route path="register" element={<RegisterPage />} />
         <Route path="app" element={<MainLayout />}>
           <Route index element={<LiveCamPage />} />
-          <Route path="users" element={<UsersPage/>}/>
+          <Route path="users" element={<UsersPage />} />
           <Route path="laporan" element={<ReportPage />} />
           {/* <Route path="websoket" element={<WebSocketDemo />} /> */}
           <Route path="test" element={<div>test view</div>} />
